Move journal/entry to journal/manager to preserve commit history
