+++ opengever/journal/manager.py
